# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ABC LTDA",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
        "/products": {
            "get": {
                "tags": [
                    "Show Products"
                ],
                "summary": "Display all products",
                "description": "Display all products",
                "operationId": "display_products",
                "responses": {
                    "200": {
                        "description": "products show successfully"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Store Products"
                ],
                "summary": "Register a new products",
                "description": "Register a new products",
                "operationId": "new_products",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text"
                                    },
                                    "description": {
                                        "type": "text"
                                    },
                                    "price": {
                                        "type": "decimal"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "name": "Celular 1",
                                    "price": 2.3,
                                    "description": "Lorem ipsum dolor sit amet"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/sales": {
            "get": {
                "tags": [
                    "Show Sales"
                ],
                "summary": "Display all sales",
                "description": "Display all sales",
                "operationId": "display_sales",
                "responses": {
                    "200": {
                        "description": "sales show successfully"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Store Sales"
                ],
                "summary": "Register a new sales",
                "description": "Register a new sales",
                "operationId": "new_sales",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": [
                                    {
                                        "tb_client_id": 1,
                                        "tb_product_id": 2,
                                        "price": 8.5,
                                        "quantity": 10
                                    }
                                ]
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/sales/{id}": {
            "get": {
                "tags": [
                    "Show Sale"
                ],
                "summary": "Displaing one sale from id the client",
                "description": "Displaing one sale from id the client",
                "operationId": "sale_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Client id",
                        "required": true
                    }
                ],
                "responses": {
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Cancel a sale"
                ],
                "summary": "Cancel a sale",
                "description": "Cancel a sale",
                "operationId": "sale_delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Sale id",
                        "required": true
                    }
                ],
                "responses": {
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    }
}